requirements-space.txt
streamlit==1.41.0
faiss-cpu==1.7.4
sentence-transformers==3.0.1
torch==2.3.0
numpy==1.26.4
huggingface_hub
python-dotenv==1.2.2